# Improve []byte to string conversion performance on NodeJS