Using `Buffer.from(array).toString("binary")` gives a 13x performance boost when doing `[]byte` to `string` conversion.

`Buffer` is only supported on NodeJS AFAIK (since v5.10.0). If `Buffer` is not supported it falls back to default implementation.

### Benchmark
The benchmark function was:
```go
var str string

func BenchmarkBytesToString(b *testing.B) {
	x := []byte("abcdefghijklmnopqrstuvwxyz")
	for i := 0; i < 15; i++ {
		x = append(x, x...)
	}
	b.SetBytes(int64(len(x)))
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		str = string(x)
	}
}
```

```
benchmark                  old ns/op     new ns/op     delta
BenchmarkBytesToString     35366666      2734000       -92.27%

benchmark                  old MB/s     new MB/s     speedup
BenchmarkBytesToString     24.09        311.62       12.94x
```