@flimzy 2017-03-24
I noticed the same behavior while working on #580, but didn't consider it worth creating a PR for three reasons:

1. Node.js is supported primarily for testing, and testing code that's different than that run in the browser seems like a bad idea. Perhaps there's a use case for targeting Node.js as a primary target though? 
2. I didn't want to increase build sizes for an optimization that would, by the previous definition, help 0% of the target environments. 
3. Buffer polyfills exist, and I wouldn't want to hurt browser performance in a case where a Buffer polyfill was loaded out of necessity (to run a node-centric module). 

Just my thoughts.