@theclapp 2017-03-24
> Node.js is supported primarily for testing

FWIW, I didn't realize this.  I figured GopherJS was for anybody that preferred Go over JS but had to run on a JS-only platform.  Just my own $0.02.