@flimzy 2017-03-24
Quoting from the top of the README:

> GopherJS compiles Go code (golang.org) to pure JavaScript code. Its main purpose is to give you the opportunity to write front-end code in Go which will still run in all browsers.

Now I don't think that has to exclude running Go code in Node.js, but a number of fundamental design decisions do seem to preclude running meaningful production code in Node.js. In particular, the lack of true syscall support in Node.js, and limited emulation of syscalls, would be blockers for all but the most trivial use cases, I suspect.

But that's not my call, so I'm only offering my observations for perspective and consideration.