@neelance 2017-03-25
Yes, I currently don't consider Node.js a primary target. I don't see any scenario where you would need it. Just run plain Go instead.