@Mikayex 2017-03-27
About the code size increase, it is only 127 bytes minified and 32 bytes minified/gzipped.

Anyway, I understand your points and close this PR.
I'll try to find other optimizations, supported in browsers this time.